Move CBT metadata and snapshots to ocs_ci/ocs/resources
Address fbalak's review feedback to move cbt_metadata and snapshot
functionality from ocs_ci/helpers to ocs_ci/ocs/resources, which is
the standard location for OCS resource-related utilities.

Also rename snapshot_helpers.py to snapshots.py to match the naming
convention in ocs_ci/ocs/resources/ (e.g., pod.py, pvc.py, not
pod_helpers.py).

Changes:
- Move cbt_metadata.py: helpers → ocs/resources
- Move and rename snapshot_helpers.py → ocs/resources/snapshots.py
- Update module docstring and circular import comment
- Update imports in test_rbd_cbt_metadata.py
