+++ ocs_ci/ocs/resources/cbt_metadata.py
